Issue #7: Adding Code Style, Tests and Code Coverage in package.json
This commit adds all the needed settings for the code coverage and
linting needed for the tests.

* Pre Tests: Run Code Style and JSHint to verify code problems.
* Test: Instruments the code with Istanbul and Runs the Mocha tests
* Post Tests: Generate more reports for coveralls and others

Just run "npm test" and all the phases will be executed in the order
described above.

The code coverage reports will be displayed in the shell as well as HTML
reports will be created under the coverage/ directory.

*	new file:   .jscsrc
*	new file:   .jshintignore
*	modified:   .jshintrc
- Used by the Pre Tests to verify code standards, style, lint.

*	modified:   package.json
- Adding the scripts for each of the phases.