Issue #7, Issue #6: Moving the CLI binary to lib and cleaning test cases
This commit moves the command binary to the lib directory. This way, all
the source code can be verified using the lint and code quality. The
code was cleaned using the same lint rules.

In addition, Adding the cli to the posttest script, so that it creates a
badge of the project itself, even though it is not needed.

The test cases are not clean as the badge file is created under the test
dir and removed after the test finishes.

Coverage is at around 90% overall.

*	deleted:    command.js
*	new file:   lib/cli.js
- Moving the command as cli
- Formatting the code according to the standards.

*	modified:   package.json
- Adding the binary call to generate the badge.

*	modified:   test/index-test.js
- Removing the generated coverage.svg file.

*	modified:   README.md
- Adding the coveralls badge