Issue #2: Updating the npm secure value from travis for the user
The Travis deploy task might be failing because of the error in the user
accounts in my ~/.npmrc. As documented at
http://docs.travis-ci.com/user/deployment/npm/, I used the command to
get the secure value for the token, but the problem was that I got for a
different user.

This commit adds the correct one. I have published 1.0.0 manually, but
moving forward, this must be from Travis. The command expects to paste
the value from the ~/.npmrc that was copied before running it.

$ travis encrypt --add deploy.api_key
Reading from stdin, press Ctrl+D when done
******566*3**5******54*3*******