Fix #17 + Enhacements: Replacing http-request with https
As the http-request dependency uses a compiled library, this was too
mcuh for a simple download. Replacing it with the https library instead.
This is due to the fact that the correct URL to download the badges is
https and not http, and the library does not follow 301.

This fixes #17 as the URL was producing wrong outputs of the badges.

*	new file:   .npmrc
- Adding the logs to be produced during install and save to be ~

*	modified:   lib/downloader.js
- Replacing http-request with the "https" library.
- Refactoring the error handling properly.

*	modified:   lib/index.js
- The URL should be 301, as it was failing with redirects when migrated.

*	modified:   package.json
- Removing the dependency.

*	modified:   test/downloader-test.js
*	modified:   test/index-test.js
*	modified:   test/istanbulReportParser-test.js
- Changing URLs for downloads to be https
- Increasing timeout when downlads happen.