Adding support for running local shields host
This commit adds support for retrieving the badge from a different host
other than the one at shields.io. They recently changed to use https
instead, which broke the request on node 0.10. The "https" module is
broken and so, the only workaround was to use http on a locally-deployed
host. Node 0.12+ might work as is.

For this reason, exposing the use of shields can work properly for those
behind the firewall/HTTP_PROXY. To bootstrap your own shields server,
just use docker and build it as follows:

1. git clone https://github.com/badges/shields
- Pull the latest
2. "echo '{}' > secret.json"
- Create the secrets for the run
3. docker build --build-arg HTTP_PROXY=$HTTP_PROXY -t badges .
- Build the docker image called "badges"
4. docker run --name badge-service -d 80:80 -e INFOSITE=http://YOUR_HOST badges
- Run the application on port 80 pointing to your hostname.
5. Badges can be retrieved by calling the following:
- YOUR_HOST/badge/style-plastic-green.svg?style=plastic

All the styles are at https://shields.io

*	modified:   lib/downloader.js
- Adjusting the client to either HTTP or HTTPS depending on the protocol

*	modified:   lib/index.js
- Collecting the shields server from API, Environment variable or
  default.